fix security errors and remove redundant class
