+++ pom.xml
@@ -15,8 +15,6 @@
 	<description>The home for writers, journalers, and others</description>
 	<properties>
 		<java.version>17</java.version>
-		<spring-security.version>5.8.1</spring-security.version>
-		<spring.version>5.8.1</spring.version>
 	</properties>
 	<dependencies>
 		<dependency>
